# Optimize utility functions