--- a/src/core/math/arrayZipMap.ts
+++ b/src/core/math/arrayZipMap.ts
@@ -1,24 +1,34 @@
-import * as R from 'ramda';
-
 type Array<T> = T[];
 
 export function arrayMap<T1, TR>(
     a1: Array<T1>,
     cb: (d1: T1, i: number) => TR
-) {
-    const mapArray = R.addIndex<T1, TR>(R.map);
+): Array<TR> {
+    const _a1_ = a1.length;
 
-    return mapArray((iValue, i) => cb(iValue, i), a1);
+    const res: Array<TR> = [];
+
+    for (let i = 0; i < _a1_; ++i) {
+        res.push(cb(a1[i], i));
+    }
+
+    return res;
 }
 
 export function arrayMap2<T1, T2, TR>(
     a1: Array<T1>,
     a2: Array<T2>,
     cb: (d1: T1, d2: T2, i: number) => TR
-) {
-    const mapArray = R.addIndex<T1, TR>(R.map);
+): Array<TR> {
+    const _a1_ = a1.length;
 
-    return mapArray((iValue, i) => cb(iValue, a2[i], i), a1);
+    const res: Array<TR> = [];
+
+    for (let i = 0; i < _a1_; ++i) {
+        res.push(cb(a1[i], a2[i], i));
+    }
+
+    return res;
 }
 
 export function arrayMap3<T1, T2, T3, TR>(
@@ -26,11 +36,16 @@ export function arrayMap3<T1, T2, T3, TR>(
     a2: Array<T2>,
     a3: Array<T3>,
     cb: (d1: T1, d2: T2, d3: T3, i: number) => TR
-) {
-    const mapArray = R.addIndex<T1, TR>(R.map);
+): Array<TR> {
+    const _a1_ = a1.length;
 
-    return mapArray((iValue, i) => cb(iValue, a2[i], a3[i], i), a1);
+    const res: Array<TR> = [];
 
+    for (let i = 0; i < _a1_; ++i) {
+        res.push(cb(a1[i], a2[i], a3[i], i));
+    }
+
+    return res;
 }
 
 export function arrayMapN<TR>(
@@ -40,7 +55,13 @@ export function arrayMapN<TR>(
     const a1 = arrays.slice(0, 1);
     const as = arrays.slice(1);
 
-    const mapArray = R.addIndex<any, TR>(R.map);
+    const _a1_ = a1.length;
 
-    return mapArray((iValue, i) => cb(i, [iValue, ...as.map(a => a[i])]), a1);
+    const res: Array<TR> = [];
+
+    for (let i = 0; i < _a1_; ++i) {
+        res.push(cb(i, a1[i], ...as.map(a => a[i])));
+    }
+
+    return res;
 }