--- a/src/core/math/matrixZipMap.ts
+++ b/src/core/math/matrixZipMap.ts
@@ -1,65 +1,98 @@
-import * as R from 'ramda';
-
 type Matrix<T> = T[][];
 
-export const shallowClone: <T>(
+export function shallowClone<T>(
     m: Matrix<T>
-) => Matrix<T> = R.map(row => row.slice(0));
+): Matrix<T> {
+    const res = [];
 
-export const traverse2 = <T1, T2, TR>(
+    for (let i = 0; i < m.length; ++i) {
+        res.push(m[i].slice(0));
+    }
+
+    return res;
+}
+
+export function traverse2<T1, T2>(
+    a1: T1[],
+    a2: T2[],
+    fn: (d1: T1, d2: T2, i1: number, i2: number) => void
+) {
+    const _a1_ = a1.length;
+    const _a2_ = a2.length;
+
+    for (let i1 = 0; i1 < _a1_; ++i1) {
+        for (let i2 = 0; i2 < _a2_; ++i2) {
+            fn(a1[i1], a2[i2], i1, i2);
+        }
+    }
+}
+
+export function traverseMap2<T1, T2, TR>(
     a1: T1[],
     a2: T2[],
     fn: (d1: T1, d2: T2, i1: number, i2: number) => TR
-) => R.addIndex<T1>(R.forEach)((d1, i1) =>
-    R.addIndex<T2>(R.forEach)((d2, i2) =>
-        fn(d1, d2, i1, i2),
-        a2),
-    a1);
+): Matrix<TR> {
+    const _a1_ = a1.length;
+    const _a2_ = a2.length;
 
-export const traverseMap2 = <T1, T2, TR>(
-    a1: T1[],
-    a2: T2[],
-    fn: (d1: T1, d2: T2, i1: number, i2: number) => TR
-): Matrix<TR> => R.addIndex<T1, TR[]>(R.map)((d1, i1) =>
-    R.addIndex<T2, TR>(R.map)((d2, i2) =>
-        fn(d1, d2, i1, i2),
-        a2),
-    a1);
+    const res: Matrix<TR> = [];
+
+    for (let i1 = 0; i1 < _a1_; ++i1) {
+        const row = [];
+        for (let i2 = 0; i2 < _a2_; ++i2) {
+            row.push(fn(a1[i1], a2[i2], i1, i2));
+        }
+
+        res.push(row);
+    }
+
+    return res;
+}
 
 export function matrixMap<T1, TR>(
     m1: Matrix<T1>,
     cb: (d1: T1, i: number, j: number) => TR
 ) {
-    const mapMatrix = R.addIndex<T1[], TR[]>(R.map);
-    const mapRow = R.addIndex<T1, TR>(R.map);
+    const res: Matrix<TR> = [];
 
-    return mapMatrix((iRow, i) =>
-        mapRow(
-            (ijValue, j) => cb(ijValue, i, j),
-            iRow
-        ), m1
-    );
+    for (let i = 0; i < m1.length; ++i) {
+        let row = [];
+        for (let j = 0; j < m1[i].length; ++j) {
+            row.push(cb(
+                m1[i][j],
+                i,
+                j
+            ));
+        }
+
+        res.push(row);
+    }
+
+    return res;
 }
 
 export function matrixMap2<T1, T2, TR>(
     m1: Matrix<T1>,
     m2: Matrix<T2> | undefined,
     cb: (d1: T1, d2: T2 | undefined, i: number, j: number) => TR
-) {
-    const mapMatrix = R.addIndex<T1[], TR[]>(R.map);
-    const mapRow = R.addIndex<T1, TR>(R.map);
+): Matrix<TR> {
+    const res: Matrix<TR> = [];
 
-    return mapMatrix((iRow, i) =>
-        mapRow(
-            (ijValue, j) => cb(
-                ijValue,
+    for (let i = 0; i < m1.length; ++i) {
+        let row = [];
+        for (let j = 0; j < m1[i].length; ++j) {
+            row.push(cb(
+                m1[i][j],
                 m2 ? m2[i][j] : undefined,
                 i,
                 j
-            ),
-            iRow
-        ), m1
-    );
+            ));
+        }
+
+        res.push(row);
+    }
+
+    return res;
 }
 
 export function matrixMap3<T1, T2, T3, TR>(
@@ -67,22 +100,25 @@ export function matrixMap3<T1, T2, T3, TR>(
     m2: Matrix<T2> | undefined,
     m3: Matrix<T3> | undefined,
     cb: (d1: T1, d2: T2 | undefined, d3: T3 | undefined, i: number, j: number) => TR
-) {
-    const mapMatrix = R.addIndex<T1[], TR[]>(R.map);
-    const mapRow = R.addIndex<T1, TR>(R.map);
+): Matrix<TR> {
+    const res: Matrix<TR> = [];
 
-    return mapMatrix((iRow, i) =>
-        mapRow(
-            (ijValue, j) => cb(
-                ijValue,
+    for (let i = 0; i < m1.length; ++i) {
+        let row = [];
+        for (let j = 0; j < m1[i].length; ++j) {
+            row.push(cb(
+                m1[i][j],
                 m2 ? m2[i][j] : undefined,
                 m3 ? m3[i][j] : undefined,
                 i,
                 j
-            ),
-            iRow
-        ), m1
-    );
+            ));
+        }
+
+        res.push(row);
+    }
+
+    return res;
 }
 
 export function matrixMap4<T1, T2, T3, T4, TR>(
@@ -91,23 +127,26 @@ export function matrixMap4<T1, T2, T3, T4, TR>(
     m3: Matrix<T3> | undefined,
     m4: Matrix<T4> | undefined,
     cb: (d1: T1, d2: T2 | undefined, d3: T3 | undefined, d4: T4 | undefined, i: number, j: number) => TR
-) {
-    const mapMatrix = R.addIndex<T1[], TR[]>(R.map);
-    const mapRow = R.addIndex<T1, TR>(R.map);
+): Matrix<TR> {
+    const res: Matrix<TR> = [];
 
-    return mapMatrix((iRow, i) =>
-        mapRow(
-            (ijValue, j) => cb(
-                ijValue,
+    for (let i = 0; i < m1.length; ++i) {
+        let row = [];
+        for (let j = 0; j < m1[i].length; ++j) {
+            row.push(cb(
+                m1[i][j],
                 m2 ? m2[i][j] : undefined,
                 m3 ? m3[i][j] : undefined,
                 m4 ? m4[i][j] : undefined,
                 i,
                 j
-            ),
-            iRow
-        ), m1
-    );
+            ));
+        }
+
+        res.push(row);
+    }
+
+    return res;
 }
 
 export function matrixMapN<TR>(
@@ -115,21 +154,21 @@ export function matrixMapN<TR>(
     m1: Matrix<any>,
     ...matrices: (any[][] | undefined)[]
 ) {
-    const mapMatrix = R.addIndex<any[], TR[]>(R.map);
-    const mapRow = R.addIndex<any, TR>(R.map);
+    const res: Matrix<TR> = [];
 
-    return mapMatrix((iRow, i) =>
-        mapRow(
-            (ijValue, j) => cb(
+    for (let i = 0; i < m1.length; ++i) {
+        let row = [];
+        for (let j = 0; j < m1[i].length; ++j) {
+            row.push(cb(
+                m1[i][j],
                 i,
                 j,
-                [
-                    ijValue,
-                    m1[i][j],
                 ...matrices.map(m => m ? m[i][j] : undefined)
-                ]
-            ),
-            iRow
-        ), m1
-    );
+            ));
+        }
+
+        res.push(row);
+    }
+
+    return res;
 }