--- a/src/dash-table/components/Table/index.tsx
+++ b/src/dash-table/components/Table/index.tsx
@@ -84,7 +84,7 @@ export default class Table extends Component<SanitizedAndDerivedProps, Standalon
         return R.any(key =>
             !DERIVED_REGEX.test(key) && props[key] !== nextProps[key],
             R.keysIn(props)
-        ) || !isEqual(state, nextState, false);
+        ) || !isEqual(state, nextState);
     }
 
     render() {