This does 2 things:

1. Optimize the utility functions (mostly removing Ramda for/map/addIndex in favor of native `for` loops)
This drops average render time from ~50ms down to ~35ms.

2. Optimize the previously optimized edges & styles calculations (https://github.com/plotly/dash-table/pull/465). This is done by removing Ramda where possible and through some other optimizations.
This drops the average render time from ~35ms to ~23ms.

Reflow and react internals remain unchanged with ~100ms and ~50ms per render respectively while table `render` itself is slashed in half. This only represents a ~10% improvement.

At this point ~45% (~10-10.5ms) of the table's `render` time is accounted for by cache busting / checks on memoized functions.